+++ include/featherdoc.hpp
@@ -244,6 +244,8 @@ class Table {
                               std::size_t column_count = 1U);
     Table insert_table_after(std::size_t row_count = 1U,
                              std::size_t column_count = 1U);
+    Table insert_table_like_before();
+    Table insert_table_like_after();
     TableRow append_row(std::size_t cell_count = 1U);
 };